+++ wasm-bridge/bridge.ts
@@ -3,7 +3,7 @@
 
 import type { BridgeBase, InitFn, InitPromise, WasmBaseInstance, WasmImports } from './types.ts';
 
-import { stringToU8a, u8aToString } from 'https://deno.land/x/polkadot@0.2.36/util/mod.ts';
+import { stringToU8a, u8aToString } from 'https://deno.land/x/polkadot/util/mod.ts';
 
 import { Wbg } from './wbg.ts';
 
@@ -17,15 +17,16 @@ import { Wbg } from './wbg.ts';
  * the native environment, providing all the plumbing needed for the Wbg classes.
  */
 export class Bridge<C extends WasmBaseInstance> implements BridgeBase<C> {
+  readonly #createWasm: InitFn<C>;
+  readonly #heap: unknown[];
+  readonly #wbg: WasmImports;
+
   #cachegetInt32: Int32Array | null;
   #cachegetUint8: Uint8Array | null;
-  #createWasm: InitFn<C>;
-  #heap: unknown[];
   #heapNext: number;
   #wasm: C | null;
   #wasmError: string | null;
   #wasmPromise: InitPromise<C> | null;
-  #wbg: WasmImports;
   #type: 'asm' | 'wasm' | 'none';
 
   constructor (createWasm: InitFn<C>) {