+++ wasm-crypto-init/wasm.ts
@@ -1,9 +1,9 @@
 
-import type { InitFn } from 'https://deno.land/x/polkadot@0.2.36/wasm-bridge/types.ts';
+import type { InitFn } from 'https://deno.land/x/polkadot/wasm-bridge/types.ts';
 import type { WasmCryptoInstance } from './types.ts';
 
-import { createWasmFn } from 'https://deno.land/x/polkadot@0.2.36/wasm-bridge/mod.ts';
-import { wasmBytes } from 'https://deno.land/x/polkadot@0.2.36/wasm-crypto-wasm/mod.ts';
+import { createWasmFn } from 'https://deno.land/x/polkadot/wasm-bridge/mod.ts';
+import { wasmBytes } from 'https://deno.land/x/polkadot/wasm-crypto-wasm/mod.ts';
 
 export { packageInfo } from './packageInfo.ts';